Use bytes for x

+++ demos/test_one_sprite.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    uint16_t x;
+    uint8_t x;
     uint8_t y;
     uint8_t ch;
